[fix] Knowledge: fix wrong metadata and filters (#4284)
## Summary

Fixes #4258 
This is a regression from #3784 - the for loop was removed but the
`filters` argument has to be specified per doc, not globally.

## Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [x] Code complies with style guidelines
- [x] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [x] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [x] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

I guess a more efficient and thorough solution would be to change the
`insert()` and `upsert()` methods to receive a list of filters, but it
would also be a breaking change.

Co-authored-by: Kaustubh <shuklakaustubh84@gmail.com>