feat: Release 1.8.1 (#4354)
# Changelog

## New Features:

- **Reasoning summaries for OpenAI**: reasoning summaries are now
available with the OpenAIResponses class
- **Neo4j Tools**: Added `Neo4jTools` to explore and manipulate graphs
in Neo4j.

## Improvements:

- **Update E2B tool to use Sandbox.create:** Update the E2BTools class
to use Sandbox.create() from the latest e2b_code_interpreter SDK,
replacing the deprecated Sandbox() initialization. No other
functionality changed.
- **Teams Timezone Identifier:** Added `timezone_identifier` to `Team`
to match `Agent` functionality.

## Bug Fixes:

- **Media artifacts validation error:** `from_artifact()` of `Image` was
missing content in its return along with url, so if the generated image
is not a url but a encoded content this would fail, or vice-versa.
- **Wrong metadata and filters:** Fixed issues where `filters` was not
correctly applied on all documents when inserting into the knowledge
base.

---------

Co-authored-by: Dirk Brand <51947788+dirkbrnd@users.noreply.github.com>
Co-authored-by: Dirk Brand <dirkbrnd@gmail.com>