feat: Add example to build Agentic RAG without APIs (#2397)
Local stack:
- Retriever: using Langchain Knowledge base: Qdrant FastEmbed and
VectorStore
- Agentic workflow using Agno and Ollama

- **Impact on metrics**: (If applicable) Describe changes in any metrics
or performance benchmarks.

I work with private startups or companies that prefer to execute open
source LLMs and embeddings without APIs. this cookbook is one such
contribution. As people prefer Langchain, and thanks to Agno, this
example is to those who wants to build Agents but with their knowledge
of Langchain.

## Checklist

- [x] Adherence to standards: Code complies with Agno’s style guidelines
and best practices.
- [x] Formatting and validation: You have run `./scripts/format.sh` and
`./scripts/validate.sh` to ensure code is formatted and linted.
- [ ] Self-review completed: A thorough review has been performed by the
contributor(s).
- [ ] Documentation: Docstrings and comments have been added or updated
for any complex logic.
- [x] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable).
- [x] Tested in a clean environment: Changes have been tested in a clean
environment to confirm expected behavior.
- [ ] Tests (optional): Tests have been added or updated to cover any
new or changed functionality.

---------

Co-authored-by: Mustafa Esoofally <coolmusta@gmail.com>