fix: improve functionality of send_media_to_model flag (#4683)
## Summary

fixes: https://github.com/agno-agi/agno/issues/4670

we currently use the `send_media_to_model` flag for the case where the
input image to agent is to be accessed by a tool and should not be
available to the model.

In this usecase if we use a tool to generate an image (custom or dalle)
and using a non multimodal model, so in this case case we append the
tool generated media now as a user message to model, we can also use the
same `send_media_to_model` to decide to not send to model, otherwise the
non multimodal model throws error.

## Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [x] Code complies with style guidelines
- [x] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [x] Self-review completed
- [x] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [x] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

Add any important context (deployment instructions, screenshots,
security considerations, etc.)