chore: Release 2.0.7 (#4679)
# Changelog

## New Features:

- **LlamaCpp Model:** Added a new LlamaCpp Model class

## Improvements:

- **AgentOS CORS Middleware**: AgentOS now automatically merges existing
CORSMiddleware allow_origin lists with what is required for the AgentOS
UI to function.
- **Update create_agent_run and File.valid_mime_types:**
1. Couldn’t get some info about files injected in tools, e.g., filename,
content type, etc. So added it in function create_agent_run.
2.
add `application/json` and `application/vnd.openxmlformats-officedocument.wordprocessingml.document` in
`File.valid_mime_types`

## Bug Fixes:

- **`input_schema` in UI breaking workflows:** for
[[this](https://gist.github.com/ysolanky/54b5361be9908a22f5b93149bd0b7acc)](https://gist.github.com/ysolanky/54b5361be9908a22f5b93149bd0b7acc)
type of workflow config where the steps are not wrapped in `Step`class,
the UI does not show the `input_schema` form for them.
- **Improve functionality of send_media_to_model flag:**
- we currently use the `send_media_to_model` flag for the case where the
input image to agent is to be accessed by a tool and should not be
available to the model.
- In this usecase if we use a tool to generate an image (custom or
DallE) and using a non-multimodal model, so in this case case we append
the tool generated media now as a user message to model, we can also use
the same `send_media_to_model` to decide to not send to model, otherwise
the non multimodal model throws error.