chore: Release 2.0.8 (#4710)
# Changelog

## New Features:

- **CometAPI**: Added support for
[CometAPI](https://api.cometapi.com/login) as a model provider.

## Improvements:

- **Access dependencies on tool:** `dependencies` are now available on
custom tools as a built-in argument. See an
[example](https://docs.agno.com/examples/concepts/agent/dependencies/access_dependencies_in_tool).
- V1 → V2 Migration Script: Updated the migration script by introducing
batching. This makes it significantly more performant for large
databases.
- Support multiple text_contents on `add_contents()` in Knowledge.
- **MultiMCPTools:** Added `allow_partial_failure` flag to
`MultiMCPTools` toolkit

## Bug Fixes:

- **Dependencies in Teams**: Ensure dependencies are correctly passed
through to member agents to reference.
- **Tool call limit on default reasoning agent:** Ensures that a
`tool_call_limit` can be passed to the default reasoning agent, avoiding
potential costly tool call loops.
- **Fix string encoding: ****Addresses issue where string encoding broke
when adding Chinese string content. Encoding now reliably works via
utf-8 encoding.
- **Fix `run_input` not being populated in session runs endpoint:**
The `/sessions/<session_id>/runs` endpoint does not
return `run_input` in case of error as it references it
from `step_executor_runs` and its empty if the workflow hasn't run yet,
so in this case we fallback to directly get it from `input`.
- **Fix State Precedence for Agents/Teams**: Fixed issues related to
state from the database that would override state passed on `agent.run`.
- **Fix early stop in workflow to return correct last output:** early
stop functionality in workflow was not updating the last output properly
for non stream cases
- **Fixed `OpenAIResponses` to remove duplicate messages for reasoning
models:**
- The `OpenAIResponses` API has the (default enabled) feature that they
store the conversation server-side and we send
a `previous_response_id` with any followup questions.
- Agno's `OpenAIResponses` class implements this feature by looking
through the message history, finding the most recent assistant response,
and using that message ID as `previous_response_id`. This part works.
However, we were then not deleting the prior messages from the request
it sends to OpenAI. As a result OpenAI concatenates its own message
history with the message history sent by Agno, resulting in an
exponential number of repeated messages as the conversation continues.

---------

Co-authored-by: Dirk Brand <dirkbrnd@gmail.com>
Co-authored-by: Dirk Brand <51947788+dirkbrnd@users.noreply.github.com>
Co-authored-by: Kaustubh <shuklakaustubh84@gmail.com>