chore: update parallel and custom function step events streaming logic (#4954)
## Summary

- from sdk we pass workflow_context in case of streaming (`step.py`)
when we do `self.active_executor.run` . This workflow context consists
of things like step id, step name, step index etc which is then put in
RunOutputEvent used by FE.
- Now in case of a custom function step, which does `agent.run()` inside
it. this agent run is completely unanimous and not in the flow of our
execution. like the user decides to trigger it. and we cant directly
inject that workflow_context into this run, so - in case of this step
even if we yield events the FE wont know about which step it belongs to
because it doesn't have access to that `step_id` etc info.

Now we inspect all events that are yielded during the workflow run, and
set the workflow context variables on them before yielding upwards.

- Also updates parallel step event streaming logic to immediately yield
events as they come so that FE can catch and display in real time.
   -  Old approach - collect all, yield at end
- New approach - stream immediately (used queues as we cant yield
directly from async running tasks)

## Type of change

- [ ] Bug fix
- [ ] New feature
- [ ] Breaking change
- [x] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [ ] Code complies with style guidelines
- [ ] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [ ] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [ ] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

Add any important context (deployment instructions, screenshots,
security considerations, etc.)