chore: v2.2.4 (#5245)
# Changelog

## New Features:

- **Granular control of history messages**: Add `num_history_messages`
to control how many messages to consider when getting them from history.
- **Update AgentOS instances from lifespan functions**: You can know
gain access to the contextual AgentOS instance in lifespan functions.
This is useful to update anything inside the AgentOS instance after it
has already been initialized and run.

## Improvements:

- **Handle media in dict inputs**: Media instances are now supported
when providing the input for a run as a dictionary, instead of as a list
of Messages.
- **Client persistence for all models**: All model implementations now
cache and reuse their HTTP clients when possible.

## Bug Fixes:

- **`OllamaEmbedder` not respecting custom dimensions parameter:**
Fixed `OllamaEmbedder` to properly pass the `dimensions` parameter to
the Ollama API during embedding generation.
- **Metrics calculation**: Fix a problem when persisting metrics
calculations when using a model with `:` in its name.
- **Agentic memory deletion**: Fix a bug related to user memory deletion
when using agentic memory management.