chore: v2.2.12 (#5420)
# Changelog

## New Features:

- **Knowledge Search Filter Expressions:** Added a new metadata-based
Filter DSL (EQ, IN, GT/LT, NOT, AND, OR) enabling targeted Knowledge
searches across. Support for PGVector only in this release, support for
more VectorDB’s to follow. Added ew cookbook examples showing practical
usage.

## Bug Fixes:

- **Gemini Parse of Any in Tool Parameters**: Fixed issues where tools
with pydantic objects in the parameters with `any` as a type was not
parsed correctly for Gemini. This affected Agentic Knowledge Filtering,
which is also fixed.