refactor: Remove unneeded code from AWS Claude model. This enables th… (#5489)
…e AWS version of the Claude model to use all features of the general
Claude model.

## Summary

Some parameters of the base Claude class such as `betas` were being
ignored because overriding functions in the AWS child class were not
aware of these parameters. This change removes all the unneeded
overrides in the AWS child class which allows parameters like `betas` to
function the same in both versions of the model.

This also removes a bunch of duplication in the AWS Claude class which
no longer needs to be maintained.

## Type of change

- [ ] Bug fix
- [ ] New feature
- [ ] Breaking change
- [X] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [X] Code complies with style guidelines
- [X] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [X] Self-review completed
- [X] Documentation updated (comments, docstrings)
- [X] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [X] Tested in clean environment
- [X] Tests added/updated (if applicable)

---

## Additional Notes

Add any important context (deployment instructions, screenshots,
security considerations, etc.)

Co-authored-by: Dirk Brand <dirkbrnd@gmail.com>