chore: Release 2.3.13 (#5722)
# Changelog

## New Features:

- **AgentOS Role-Based Access Control (RBAC)** → Introducing RBAC for
AgentOS (see the docs). This allows you to either automatically or
manually configure JWT-based authorization checks in AgentOS with the
following benefits:
- All traffic to your AgentOS would require a signed JWT token, with the
correct user permission scopes.
- Per-endpoint authorization → Based on the scopes inside the JWT token,
access is granted or denied to each endpoint as you choose. The required
scopes are configurable as well.
- Per-agent resource control → By setting specific scopes like
`agents:my-agent:read` you can control which agents can be run (i.e.
when calling `POST /agents/my-agent/runs` ) or which agents are returned
on read (i.e. when calling `GET /agents` or `GET /agents/{id}`). This
allows for full control over which users can use and see which agents,
and this extends to AgentOS. The same also applies to `Team` and
`Workflow`.
    - AgentOS UI compatibility to be released in the coming days.
    - **Note**: On the `JWTMiddleware` class:
- `secret_key` has been deprecated (although still supported) and it is
recommended to supply `verification_keys=[...]` going forward.
- `algorithm` default value changed from `HS256` to `RS256` to make
asymmetric keys the default (better suited for production cases).